+++ Tests/MapboxMapsTests/Location/Puck/Mocks/MockPuckManager.swift
@@ -4,4 +4,6 @@ final class MockPuckManager: PuckManagerProtocol {
     var puckType: PuckType?
 
     var puckBearingSource: PuckBearingSource = .heading
+
+    var puckBearingEnabled: Bool = true
 }